+++ app/main.jsx
@@ -3,14 +3,15 @@ import React from 'react'
 import {Router, Route, IndexRedirect, browserHistory} from 'react-router'
 import {render} from 'react-dom'
 import {connect, Provider} from 'react-redux'
-
 import store from './store'
 import Int from './components/Int'
 import Login from './components/Login'
 import WhoAmI from './components/WhoAmI'
 import SignUpContainer from './components/SignUpContainer'
 import Game from './components/Game'
-import Modal from 'react-modal'
+// import Modal from 'react-modal'
+import ExampleApp from './components/ExampleApp'
+
 
 const App = connect(
   ({ auth }) => ({ user: auth })
@@ -31,75 +32,6 @@ const App = connect(
   }
 )
 
-const customStyles = {
-  overlay : {
-    position          : 'fixed',
-    top               : 300,
-    left              : 300,
-    right             : 300,
-    bottom            : 300,
-    backgroundColor   : 'rgba(255, 255, 255, 0.75)'
-  },
-  content : {
-    position                   : 'absolute',
-    top                        : '20px',
-    left                       : '20px',
-    right                      : '20px',
-    bottom                     : '20px',
-    border                     : '1px solid #ccc',
-    background                 : 'grey',
-    overflow                   : 'auto',
-    WebkitOverflowScrolling    : 'touch',
-    borderRadius               : '4px',
-    outline                    : 'none',
-    padding                    : '20px'
-
-  }
-};
-
-
-class ExampleApp extends React.Component{
-
-   constructor(props){
-    super(props)
-
-    this.state = {
-      modalIsOpen: false,
-      loginOrSignup: ''
-    }
-    this.openModal = this.openModal.bind(this)
-    this.closeModal = this.closeModal.bind(this)
-  }
-
-  openModal(e){
-    this.setState({modalIsOpen: true})
-    this.setState({loginOrSignup: e.target.name})
-  }
-
-  closeModal(){
-    this.setState({modalIsOpen: false})
-  }
-
-  render(){
-    return(
-      <div>
-          <button name="login" onClick={this.openModal}> Login in! </button>
-          <button name="signup" onClick={this.openModal}> Sign up! </button>
-          <Modal
-            isOpen={this.state.modalIsOpen}
-            onRequestClose={this.closeModal}
-            style={customStyles}
-            contentLabel="login Modal"
-          >
-          {
-            this.state.loginOrSignup === 'signup' ? <SignUpContainer /> : <Login />
-          }
-          </Modal>
-      </div>
-    )
-  }
-
-}
 
 export default class AppContainer extends React.Component{
   render(){
@@ -110,17 +42,6 @@ export default class AppContainer extends React.Component{
     )
   }
 }
-
-
-/*connect() => function called connectFunc
-connectFunc(({ user, children }) =>
-    <div>
-      <nav>
-        {user ? <WhoAmI/> : <Login/>}
-      </nav>
-      {children}
-    </div>) */
-
 
 render (
   <Provider store={store}>
